% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_drug_target_d2c.R
\name{build_drug_target_d2c}
\alias{build_drug_target_d2c}
\title{Build Drug2Cell matrix based on the disease atlas and drug targets.}
\usage{
build_drug_target_d2c(
  dat,
  gene_list,
  drug_target_df,
  n_bins = 25,
  ctrl_size = 50,
  seed = 42,
  out_path = "results"
)
}
\arguments{
\item{dat}{A Seurat object for the log-normalised, integrated scRNA-seq disease atlas.}

\item{gene_list}{The list of intersect genes between the atlas and the CMAP L1000 disease tissue specific references.}

\item{drug_target_df}{The data.frame of each drug's targets.}

\item{n_bins}{The number of bins determines how finely the background expression is segmented into groups of genes with similar expression patterns.}

\item{ctrl_size}{The number of genes that are randomly selected from the background (control) group to normalize the expression data for each drug target group.}

\item{seed}{The random seed for reproducibility, 42 by default.}

\item{out_path}{The path where the RDS file of the constructed Drug2Cell matrix is saved.}
}
\value{
The data.frame of the Drug2Cell matrix, whose rows are drugs and columns are cell barcodes.
}
\description{
The Drug2Cell matrix is constructed by calculating the average gene expression for each drug's target group in each cell and adjusting for background expression by subtracting baseline biases. This is an R implementation of the Drug2Cell Python pipeline designed by Kanemaru et al. (PMID: 37438528).
}
\references{
Kanemaru K. et al. (2023). \emph{Spatially resolved multiomics of human cardiac niches}. Nature.
}
