% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_drug_prom_inh.R
\name{compute_drug_prom_inh}
\alias{compute_drug_prom_inh}
\title{Estimate the promotion/inhibition effect of each drug on different cell types in the disease atlas.}
\usage{
compute_drug_prom_inh(
  dat,
  d2c_mat,
  disease,
  out_type = "weight",
  out_path = "results"
)
}
\arguments{
\item{dat}{A Seurat object for the log-normalised, integrated scRNA-seq disease atlas.}

\item{d2c_mat}{The data.frame of the Drug2Cell matrix constructed by the build_drug_target_d2c function.}

\item{disease}{The name of the investigated disease.}

\item{out_type}{The type of returned promotion/inhibition effect scores, by default "weight", as computed in Eq.(7) of the scDrugLink paper.}

\item{out_path}{The path where the csv files of adjusted p-values and weights are saved.}
}
\value{
The data.frame of promotion/inhibition effect weights (scores) for each pair of cell type (row) and drug (column).
}
\description{
Each drug's promotion/inhibition effect on each cell type is calculated by comparing scores derived from control and diseased cells using within-cell-type Wilcoxon rank-sum test, and adjusted for multiple comparisons using the Benjamini-Hochberg procedure.
}
