% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_drug_sens_res.R
\name{compute_drug_sens_res}
\alias{compute_drug_sens_res}
\title{Estimate the sensitivity/resistance effect of each drug on different cell types in the disease atlas.}
\usage{
compute_drug_sens_res(
  disease,
  perturbation_matrix_path,
  gene_info,
  drug_info,
  deg_list,
  out_path = "results"
)
}
\arguments{
\item{disease}{The name of the investigated disease.}

\item{perturbation_matrix_path}{The path of the CMAP L1000 disease tissue-specific pertuabtion signature matrix.}

\item{gene_info}{The meta data for genes in the perturbation matrix.}

\item{drug_info}{The meta data for drugs in the perturbation matrix.}

\item{deg_list}{The list of intra-cell-type DEGs identified by the get_intra_cell_type_degs function.}

\item{out_path}{The path where the csv file of the p-values and the adjusted p-values (FDRs) for each pair of drug and cell type is saved.}
}
\value{
The data.frame of the p-values and the adjusted p-values (FDRs) for statistical tests that estimate sensitivity/resistance effect for each pair of cell type (row) and drug (column).
}
\description{
This function computes the p-values and adjusted p-values for each drug-cell type comparison via reverse gene expression pattern matching and K-S test (following the Asgard pipeline), which is an essential step for estimating the sensitivity/resistance effect.
}
\references{
He B. et al. (2023). \emph{ASGARD is A Single-cell Guided Pipeline to Aid Repurposing of Drugs}. Nature Communications.
}
