% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scdruglink_score.R
\name{compute_scdruglink_score}
\alias{compute_scdruglink_score}
\title{Link the drug promotion/inhibition and drug sensitivity/resistance estimations to generate an overall therapeutic score}
\usage{
compute_scdruglink_score(
  dat,
  deg_list,
  drug_sens_res_pval,
  tissue,
  gse70138_gctx_path,
  gse92742_gctx_path,
  drug_prom_inh_weight,
  clusters = NULL,
  case = NULL,
  fda_drugs_only = TRUE,
  out_path = "results"
)
}
\arguments{
\item{dat}{A Seurat object for a log-normalised, integrated scRNA-seq disease atlas}

\item{deg_list}{The list of intra-cell-type DEGs identified by the get_intra_cell_type_degs function.}

\item{drug_sens_res_pval}{the p-values and the adjusted p-values (FDRs) for statistical tests that estimate sensitivity/resistance effect for each pair of cell type and drug.}

\item{tissue}{The name of the disease-specific tissue in CMAP L1000.}

\item{gse70138_gctx_path}{The path of the CMAP L1000 GSE70138 data.}

\item{gse92742_gctx_path}{The path of the CMAP L1000 GSE92742 data.}

\item{drug_prom_inh_weight}{The promotion/inhibition effect weights for each pair of cell type and drug.}

\item{clusters}{The specific cell types/clusters on which computation of the final drug scores is based, NULL by default for all cell types/clusters.}

\item{case}{The name of the disease in the atlas.}

\item{fda_drugs_only}{Whether only FDA approved drugs are considred for repurposing, TRUE by default. This parameter is more relevant in Asgard, as scDrugLink determines the drugs and fix them from the beginning of the pipeline.}

\item{out_path}{The path where the csv files of the final drug scores at the cell-type level and the whole-atlas levl are saved.}
}
\value{
The data.frame of the final drug scores at the whole-atlas level.
}
\description{
The final drug score is calculated by summing the promotion/inhibition-weighted sensitivity/resistance scores for each cell type, as described by Eq. (7-9) in the scDrugLink paper. This function extends the DrugScore function of Asgard to facilitate the linking mechanism proposed in the scDrugLink paper.
}
