% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intra_cell_type_degs.R
\name{get_intra_cell_type_degs}
\alias{get_intra_cell_type_degs}
\title{Identify within cell-type differentially expressed genes (DEGs) between diseased and control cells.}
\usage{
get_intra_cell_type_degs(dat, disease, control)
}
\arguments{
\item{dat}{A Seurat object for a log-normalised, integrated scRNA-seq disease atlas.}

\item{disease}{The disease term in the atlas for identifying diseased cells.}

\item{control}{The control term in the atlas for identifying control cells.}
}
\value{
A list of DEGs for each cell type.
}
\description{
This function processes each cell type individually, performs statistical testing to find markers, and stores the results (log fold change and adjusted p-values) for each cell type in a list.
}
